annoying float warning

+++ MapView/Map/RMLayerCollection.h
@@ -57,7 +57,7 @@
 
 - (void)moveToProjectedPoint: (RMProjectedPoint)aPoint;
 - (void)moveBy: (CGSize) delta;
-- (void)zoomByFactor: (float) zoomFactor near:(CGPoint) center;
+- (void)zoomByFactor: (double) zoomFactor near:(CGPoint) center;
 - (void)removeSublayer:(CALayer *)layer;
 - (void)removeSublayers:(NSArray *)layers;
 /// (guess) recompute the screen coordinates for the sublayers (map markers, paths)
